# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: K&S Solar Energy API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /site-visits/{id}/technicians:
    put:
      operationId: assignSiteVisitTechnicians
      tags: [site-visits]
      summary: Assign multiple technicians to a site visit (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignTechniciansRequest"
      responses:
        "200":
          description: Technicians assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteVisit"
        "403":
          description: Forbidden
        "404":
          description: Not found
...
  /site-visits:
    get:
      operationId: getSiteVisits
      tags: [siteVisits]
      summary: List site visits (admin all, technician assigned)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of site visits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SiteVisit'
    post:
      operationId: createSiteVisit
      tags: [siteVisits]
      summary: Create a site visit (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSiteVisitRequest'
      responses:
        '201':
          description: Created site visit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SiteVisit'
  /site-visits/{id}:
    patch:
      operationId: updateSiteVisit
      tags: [siteVisits]
      summary: Update a site visit (admin full, technician status+notes only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSiteVisitRequest'
      responses:
        '200':
          description: Updated site visit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SiteVisit'
    delete:
      operationId: deleteSiteVisit
      tags: [siteVisits]
      summary: Delete a site visit (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AssignTechniciansRequest:
      type: object
      properties:
        technicianIds:
          type: array
          items:
            type: string
      required:
        - technicianIds
...
    SiteVisit:
      type: object
      required: [id, customerName, phone, address, purpose, status, technicianIds, createdAt]
      properties:
        id:
          type: string
        customerName:
          type: string
        phone:
          type: string
        address:
          type: string
        city:
          type: string
          nullable: true
        purpose:
          type: string
        notes:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, in_progress, completed, cancelled]
        assignedTo:
          type: string
          nullable: true
        technicianName:
          type: string
          nullable: true
        technicianIds:
          type: array
          items:
            type: string
        scheduledDate:
          type: string
          nullable: true
        scheduledTime:
          type: string
          nullable: true
        technicianNotes:
          type: string
          nullable: true
        createdAt:
          type: string
    CreateSiteVisitRequest:
      type: object
      required: [customerName, phone, address, purpose]
      properties:
        customerName:
          type: string
        phone:
          type: string
        address:
          type: string
        city:
          type: string
        purpose:
          type: string
        notes:
          type: string
        assignedTo:
          type: string
        scheduledDate:
          type: string
        scheduledTime:
          type: string
    UpdateSiteVisitRequest:
      type: object
      properties:
        customerName:
          type: string
        phone:
          type: string
        address:
          type: string
        city:
          type: string
        purpose:
          type: string
        notes:
          type: string
        assignedTo:
          type: string
        scheduledDate:
          type: string
        scheduledTime:
          type: string
        status:
          type: string
        technicianNotes:
          type: string
...