# lib/api-zod/src/generated/types/siteVisit.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * K&S Solar Energy API
 * OpenAPI spec version: 0.1.0
 */
import type { SiteVisitStatus } from './siteVisitStatus';

export interface SiteVisit {
  id: string;
  customerName: string;
  phone: string;
  address: string;
  city?: string | null;
  purpose: string;
  notes?: string | null;
  status: SiteVisitStatus;
  assignedTo?: string | null;
  technicianName?: string | null;
  technicianIds: string[];
  scheduledDate?: string | null;
  scheduledTime?: string | null;
  technicianNotes?: string | null;
  createdAt: string;
}